+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - [SIL.TestUtilities] Added a Create method to TemporaryFolder that takes a TestContext
+- [SIL.Media] Added .NET 8-Windows target (does not support WindowsAudioSession)
+- [SIL.Core] Support for RobustFile.GetAccessControl in all builds
+
 
 ### Fixed